Move PageCacheRule and StandalonePageCacheFactory into io module
