+++ community/kernel/src/test/java/org/neo4j/kernel/impl/pagecache/ConfigurableStandalonePageCacheFactoryTest.java
@@ -35,6 +35,7 @@ import org.neo4j.io.fs.FileSystemAbstraction;
 import org.neo4j.io.pagecache.PageCache;
 import org.neo4j.io.pagecache.PageCursor;
 import org.neo4j.io.pagecache.PagedFile;
+import org.neo4j.io.pagecache.impl.muninn.StandalonePageCacheFactory;
 
 import static org.junit.Assert.assertTrue;
